Update unit tests for env_resources.py
