+++ libensemble/tests/unit_tests/test_resources.py
@@ -1,5 +1,6 @@
 import os
 import socket
+from libensemble.env_resources import EnvResources
 from libensemble.resources import Resources, WorkerResources, ResourcesException
 
 
@@ -39,191 +40,58 @@ def teardown_function(function):
 
 # Tests ========================================================================================
 
-# Tests for obtaining nodelist from environment variables
-
-def test_slurm_nodelist_empty():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = ""
-    exp_out = []  # empty
-    nodelist = Resources.get_slurm_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_slurm_nodelist_single():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "knl-0056"
-    exp_out = ["knl-0056"]
-    nodelist = Resources.get_slurm_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_slurm_nodelist_knl_seq():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "knl-[0009-0012]"
-    exp_out = ['knl-0009', 'knl-0010', 'knl-0011', 'knl-0012']
-    nodelist = Resources.get_slurm_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_slurm_nodelist_bdw_seq():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "bdw-[0009-0012]"
-    exp_out = ['bdw-0009', 'bdw-0010', 'bdw-0011', 'bdw-0012']
-    nodelist = Resources.get_slurm_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_slurm_nodelist_groups():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "knl-[0020-0022,0137-0139,1234]"
-    exp_out = ['knl-0020', 'knl-0021', 'knl-0022', 'knl-0137', 'knl-0138', 'knl-0139', 'knl-1234']
-    nodelist = Resources.get_slurm_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_slurm_nodelist_groups_longprefix():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "super-[000020-000022,000137-000139,001234,023456-023458]"
-    exp_out = ['super-000020', 'super-000021', 'super-000022', 'super-000137', 'super-000138', 'super-000139',
-               'super-001234', 'super-023456', 'super-023457', 'super-023458']
-    nodelist = Resources.get_slurm_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_slurm_nodelist_reverse_grp():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "knl-[0020-0022,0139-0137,1234]"
-    exp_out = ['knl-0020', 'knl-0021', 'knl-0022', 'knl-0137', 'knl-0138', 'knl-0139', 'knl-1234']
-    nodelist = Resources.get_slurm_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_cobalt_nodelist_empty():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = ""
-    exp_out = []  # empty
-    nodelist = Resources.get_cobalt_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_cobalt_nodelist_single():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "56"
-    exp_out = ["56"]
-    nodelist = Resources.get_cobalt_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_cobalt_nodelist_seq():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "9-12"
-    exp_out = ['9', '10', '11', '12']
-    nodelist = Resources.get_cobalt_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_cobalt_nodelist_groups():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "20-22,137-139,1234"
-    exp_out = ['20', '21', '22', '137', '138', '139', '1234']
-    nodelist = Resources.get_cobalt_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_cobalt_nodelist_reverse_grp():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "20-22,139-137,1234"
-    exp_out = ['20', '21', '22', '137', '138', '139', '1234']
-    nodelist = Resources.get_cobalt_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_lsf_nodelist_empty():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = ""
-    exp_out = []  # empty
-    nodelist = Resources.get_lsf_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_lsf_nodelist_single():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = 'batch5' + ' g06n02' * 42
-    exp_out = ["g06n02"]
-    nodelist = Resources.get_lsf_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_lsf_nodelist_seq():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = 'batch5' + ' g06n02' * 42 + ' h21n18' * 42
-    exp_out = ['g06n02', 'h21n18']
-    nodelist = Resources.get_lsf_nodelist(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-# These dont apply to the lsf lists as they are listed in full
-# def test_lsf_nodelist_groups():
-# def test_lsf_nodelist_reverse_grp():
-
-
-def test_lsf_nodelist_shortform_empty():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = ""
-    exp_out = []  # empty
-    nodelist = Resources.get_lsf_nodelist_frm_shortform(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_lsf_nodelist_shortform_single():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = 'batch5 1 g06n02 42'
-    exp_out = ["g06n02"]
-    nodelist = Resources.get_lsf_nodelist_frm_shortform(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
-def test_lsf_nodelist_shortform_seq():
-    os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = 'batch5 1 g06n02 42 h21n18 42'
-    exp_out = ['g06n02', 'h21n18']
-    nodelist = Resources.get_lsf_nodelist_frm_shortform(node_list_env="LIBE_RESOURCES_TEST_NODE_LIST")
-    assert nodelist == exp_out, "Nodelist returned does not match expected"
-
-
 # Tests Resources.get_global_nodelist (This requires above tests to work)
 def test_get_global_nodelist_frm_slurm():
     os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "knl-[0020-0022,0137-0139,1234]"
     exp_out = ['knl-0020', 'knl-0021', 'knl-0022', 'knl-0137', 'knl-0138', 'knl-0139', 'knl-1234']
-    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(),
-                                                    nodelist_env_slurm="LIBE_RESOURCES_TEST_NODE_LIST",
-                                                    nodelist_env_cobalt="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    env_resources = EnvResources(nodelist_env_slurm="LIBE_RESOURCES_TEST_NODE_LIST",
+                                 nodelist_env_cobalt="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(), env_resources=env_resources)
     assert global_nodelist == exp_out, "global_nodelist returned does not match expected"
 
 
 def test_get_global_nodelist_frm_cobalt():
     os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = "20-22,137-139,1234"
     exp_out = ['20', '21', '22', '137', '138', '139', '1234']
-    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(),
-                                                    nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_cobalt="LIBE_RESOURCES_TEST_NODE_LIST",
-                                                    nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    env_resources = EnvResources(nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_cobalt="LIBE_RESOURCES_TEST_NODE_LIST",
+                                 nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(), env_resources=env_resources)
     assert global_nodelist == exp_out, "global_nodelist returned does not match expected"
 
 
 def test_get_global_nodelist_frm_lsf():
     os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = 'batch5' + ' g06n02' * 42 + ' h21n18' * 42
     exp_out = ['g06n02', 'h21n18']
-    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(),
-                                                    nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_cobalt="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_lsf="LIBE_RESOURCES_TEST_NODE_LIST",
-                                                    nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    env_resources = EnvResources(nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_cobalt="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf="LIBE_RESOURCES_TEST_NODE_LIST",
+                                 nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(), env_resources=env_resources)
     assert global_nodelist == exp_out, "global_nodelist returned does not match expected"
 
 
 def test_get_global_nodelist_frm_lsf_shortform():
     os.environ["LIBE_RESOURCES_TEST_NODE_LIST"] = 'batch5 1 g06n02 42 h21n18 42'
     exp_out = ['g06n02', 'h21n18']
-    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(),
-                                                    nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_cobalt="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_lsf_shortform="LIBE_RESOURCES_TEST_NODE_LIST")
+    env_resources = EnvResources(nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_cobalt="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf_shortform="LIBE_RESOURCES_TEST_NODE_LIST")
+    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(), env_resources=env_resources)
     assert global_nodelist == exp_out, "global_nodelist returned does not match expected"
 
 
 def test_get_global_nodelist_standalone():
     mynode = socket.gethostname()
-    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(),
-                                                    nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_cobalt="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                    nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    env_resources = EnvResources(nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_cobalt="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    global_nodelist = Resources.get_global_nodelist(rundir=os.getcwd(), env_resources=env_resources)
     assert global_nodelist == [mynode], "global_nodelist returned does not match expected"
 
 
@@ -249,9 +117,11 @@ def test_get_global_nodelist_frm_wrklst_file():
     assert global_nodelist1 == exp_out, "global_nodelist returned does not match expected"
 
     # Specify env vars - should ignore
-    global_nodelist2 = Resources.get_global_nodelist(rundir=os.getcwd(),
-                                                     nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
-                                                     nodelist_env_cobalt="LIBE_RESOURCES_TEST_NODE_LIST")
+    env_resources = EnvResources(nodelist_env_slurm="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_cobalt="LIBE_RESOURCES_TEST_NODE_LIST",
+                                 nodelist_env_lsf="THIS_ENV_VARIABLE_IS_DEF_NOT_SET",
+                                 nodelist_env_lsf_shortform="THIS_ENV_VARIABLE_IS_DEF_NOT_SET")
+    global_nodelist2 = Resources.get_global_nodelist(rundir=os.getcwd(), env_resources=env_resources)
     assert global_nodelist2 == exp_out, "global_nodelist returned does not match expected"
     os.remove('worker_list')
 
@@ -515,28 +385,6 @@ def test_worker_resources():
 
 if __name__ == "__main__":
     setup_standalone_run()
-
-    test_slurm_nodelist_empty()
-    test_slurm_nodelist_single()
-    test_slurm_nodelist_knl_seq()
-    test_slurm_nodelist_bdw_seq()
-    test_slurm_nodelist_groups()
-    test_slurm_nodelist_groups_longprefix()
-    test_slurm_nodelist_reverse_grp()
-
-    test_cobalt_nodelist_empty()
-    test_cobalt_nodelist_single()
-    test_cobalt_nodelist_seq()
-    test_cobalt_nodelist_groups()
-    test_cobalt_nodelist_reverse_grp()
-
-    test_lsf_nodelist_empty()
-    test_lsf_nodelist_single()
-    test_lsf_nodelist_seq()
-
-    test_lsf_nodelist_shortform_empty()
-    test_lsf_nodelist_shortform_single()
-    test_lsf_nodelist_shortform_seq()
 
     test_get_global_nodelist_frm_slurm()
     test_get_global_nodelist_frm_cobalt()
